refactor: Move types into types directory
