+++ components/BlogPosts.tsx
@@ -1,6 +1,6 @@
 import BlogPost from "components/BlogPost";
 import GoToLink from "components/GoToLink";
-import { Post } from "lib/types";
+import { Post } from "types";
 
 interface Props {
   posts: Post[];
@@ -8,13 +8,13 @@ interface Props {
 
 export default function BlogPosts({ posts }: Props) {
   return (
-    <div className='flex flex-col space-y-6'>
-      <div className='flex flex-col space-y-8'>
+    <div className="flex flex-col space-y-6">
+      <div className="flex flex-col space-y-8">
         {posts.map((post) => (
           <BlogPost key={post._id} {...post} />
         ))}
       </div>
-      <GoToLink href='/blog' text='Go to all blog posts' />
+      <GoToLink href="/blog" text="Go to all blog posts" />
     </div>
   );
 }
