+++ layouts/blog.tsx
@@ -4,7 +4,7 @@ import Image from "next/future/image";
 
 import Container, { CustomMeta } from "components/Container";
 import ViewCounter from "components/ViewCounter";
-import { Post } from "lib/types";
+import { Post } from "types";
 
 interface Props {
   children: React.ReactNode;
@@ -20,33 +20,35 @@ export default function BlogLayout({ children, post }: Props) {
   };
   return (
     <Container customMeta={customMeta}>
-      <article className='flex flex-col items-start justify-center w-full max-w-2xl mx-auto mb-16'>
-        <h1 className='mb-4 text-3xl font-bold tracking-tight text-black md:text-5xl dark:text-white'>
+      <article className="flex flex-col items-start justify-center w-full max-w-2xl mx-auto mb-16">
+        <h1 className="mb-4 text-3xl font-bold tracking-tight text-black md:text-5xl dark:text-white">
           {post.title}
         </h1>
-        <div className='flex flex-col items-start justify-between w-full mt-2 md:flex-row md:items-center'>
-          <div className='flex items-center'>
+        <div className="flex flex-col items-start justify-between w-full mt-2 md:flex-row md:items-center">
+          <div className="flex items-center">
             <Image
-              alt='Tony Pettigrew'
+              alt="Tony Pettigrew"
               height={24}
               width={24}
-              sizes='20vw'
-              src='/avatar.jpeg'
-              className='rounded-full'
+              sizes="20vw"
+              src="/avatar.jpeg"
+              className="rounded-full"
             />
-            <p className='ml-2 text-sm text-gray-700 dark:text-gray-300'>
+            <p className="ml-2 text-sm text-gray-700 dark:text-gray-300">
               {"Tony Pettigrew / "}
               {format(parseISO(post.date), "MMMM dd, yyyy")}
             </p>
           </div>
-          <p className='mt-2 text-sm text-gray-600 dark:text-gray-400 min-w-32 md:mt-0'>
+          <p className="mt-2 text-sm text-gray-600 dark:text-gray-400 min-w-32 md:mt-0">
             {post.readingTime}
             {` • `}
-            <ViewCounter slug={post.slug} type='blog' />
+            <ViewCounter slug={post.slug} type="blog" />
           </p>
         </div>
         <Suspense fallback={null}>
-          <div className='w-full mt-4 prose dark:prose-dark max-w-none'>{children}</div>
+          <div className="w-full mt-4 prose dark:prose-dark max-w-none">
+            {children}
+          </div>
         </Suspense>
       </article>
     </Container>
