+++ types/Views.ts
@@ -1,0 +1,3 @@
+export interface Views {
+  total: string;
+}
