+++ types/index.ts
@@ -1,0 +1,6 @@
+export * from "./Post";
+export * from "./Snippet";
+export * from "./Project";
+export * from "./Spotify";
+export * from "./Github";
+export * from "./Views";
